move ref types to parser module
